+++ realm/realm-library/src/objectServer/java/io/realm/mongodb/mongo/events/OperationType.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.realm.mongodb.mongo.remote;
+package io.realm.mongodb.mongo.events;
 
 public enum OperationType {
     INSERT, DELETE, REPLACE, UPDATE, UNKNOWN;
